change directory architecture
