+++ lib/core/src/lib.rs
@@ -1,4 +1,3 @@
 mod positive_integer;
 pub mod source_code;
 pub mod token;
-mod token_prepared;
