+++ lib/scanner/src/lib.rs
@@ -1,3 +1,3 @@
-mod buffer;
+pub mod buffer;
 pub mod scanner;
 mod token_generator;
